# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MeshCore Home",
    "version": "api-1"
  },
  "paths": {
...
    "/api/setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Run Setup",
        "operationId": "run_setup_api_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedIn"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedIn"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/sessions/{session_id}": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Sign Out Session",
        "description": "Sign out one browser or app (e.g. a lost phone). Its open connections close at once.",
        "operationId": "sign_out_session_api_auth_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "operationId": "change_password_api_auth_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/username": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "Change Username",
        "operationId": "change_username_api_auth_username_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUsernameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Current Password"
          },
          "new_password": {
            "type": "string",
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest"
      },
      "ChangeUsernameRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.@-]+$",
            "title": "Username"
          },
          "current_password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Current Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "current_password"
        ],
        "title": "ChangeUsernameRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "client": {
            "type": "string",
            "enum": [
              "web",
              "ios",
              "android"
            ],
            "title": "Client",
            "default": "web"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "Me": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "home_name": {
            "type": "string",
            "title": "Home Name"
          }
        },
        "type": "object",
        "required": [
          "username",
          "home_name"
        ],
        "title": "Me"
      },
...
      "RadioConfig": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "simulated",
              "tcp",
              "hat",
              "none"
            ],
            "title": "Mode",
            "default": "simulated"
          },
          "host": {
            "type": "string",
            "title": "Host",
            "default": ""
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port",
            "default": 5000
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "default": false
          },
          "sim_interval_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Sim Interval Seconds",
            "default": 60
          }
        },
        "type": "object",
        "title": "RadioConfig"
      },
...
      "SetupRequest": {
        "properties": {
          "client": {
            "type": "string",
            "enum": [
              "web",
              "ios",
              "android"
            ],
            "title": "Client",
            "default": "web"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "setup_token": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Setup Token"
          },
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.@-]+$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "home_name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Home Name",
            "default": "Home"
          },
          "radio": {
            "$ref": "#/components/schemas/RadioConfig"
          }
        },
        "type": "object",
        "required": [
          "setup_token",
          "username",
          "password",
          "radio"
        ],
        "title": "SetupRequest"
      },
...
      "SignedIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "home_name": {
            "type": "string",
            "title": "Home Name"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "username",
          "home_name"
        ],
        "title": "SignedIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key from Settings → API keys (mch_...). See docs/API.md."
      }
    }
  },
  "security": [
    {
      "ApiKey": []
    }
  ]
}